+++ client/src/Medical-Chatbot/script.js
