+++ client/src/Medical-Chatbot/training-chatbot/intents.json
