+++ client/src/components/FloatingIcons.jsx
