+++ client/src/components/PatientDistributionChart.jsx
