--- client/src/modules/common/Medical-Chatbot/config.json
@@ -1,3 +1,0 @@
-{
-  "apiKey": "add your api key"
-}
