+++ client/src/pages/AuthForm.jsx
@@ -1,11 +1,11 @@
 import React, { useState } from 'react';
 import { FaEye, FaEyeSlash } from 'react-icons/fa';
 import { TailSpin } from 'react-loader-spinner';
-import { notify } from './notification';
-import '../../styles/Login.css';
+import { notify } from '../components/notification';
+import '../styles/Login.css';
 import { Link } from 'react-router-dom';
 import { useRecoilValue } from 'recoil'; // Import recoil
-import { mode } from '../../store/atom'; // Import dark mode atom
+import { mode } from '../store/atom'; // Import dark mode atom
 
 const AuthPage = () => {
   const [formData, setFormData] = useState({
@@ -115,44 +115,39 @@ const AuthPage = () => {
 
   return (
     <div
-      className={`login_background ${
-        dark === 'dark' ? ' text-yellow-400' : 'bg-white text-gray-900'
-      }`}
+      className={`login_background ${dark === 'dark' ? ' text-yellow-400' : 'bg-white text-gray-900'
+        }`}
     >
       <div className="auth-maindiv">
         <div
-          className={`auth-container ${
-            dark === 'dark' ? ' text-yellow-400' : 'bg-white text-gray-900'
-          }`}
+          className={`auth-container ${dark === 'dark' ? ' text-yellow-400' : 'bg-white text-gray-900'
+            }`}
         >
-          <h2 className={` ${
-              dark === 'dark'
-                ? ' text-yellow-400'
-                : 'bg-white text-gray-700'
+          <h2 className={` ${dark === 'dark'
+            ? ' text-yellow-400'
+            : 'bg-white text-gray-700'
             }`}>Login</h2>
           <form onSubmit={handleSubmit} className="auth-form">
             <div className="form-section">
-              <label className={`auth-form ${
-              dark === 'dark'
+              <label className={`auth-form ${dark === 'dark'
                 ? ' text-yellow-400'
                 : 'bg-white text-gray-700'
-            }`}>User Type:</label>
+                }`}>User Type:</label>
               <select
                 name="type"
                 value={formData.type}
                 onChange={handleChange}
-                
+
               >
                 <option value="user">User</option>
                 <option value="hospital">Hospital</option>
               </select>
             </div>
             <div className="form-section">
-              <label className={`auth-form ${
-              dark === 'dark'
+              <label className={`auth-form ${dark === 'dark'
                 ? ' text-yellow-400'
                 : 'bg-white text-gray-700'
-            }`}>Email:</label>
+                }`}>Email:</label>
               <input
                 type="email"
                 name="email"
@@ -160,9 +155,8 @@ const AuthPage = () => {
                 value={formData.email}
                 onChange={handleChange}
                 required
-                className={`${
-                  dark === 'dark' ? ' text-yellow-400' : 'bg-white text-gray-900'
-                }`}
+                className={`${dark === 'dark' ? ' text-yellow-400' : 'bg-white text-gray-900'
+                  }`}
               />
               {errors.frontend.email && (
                 <span className="error">{errors.frontend.email}</span>
@@ -173,11 +167,10 @@ const AuthPage = () => {
             </div>
 
             <div className="form-section">
-              <label className={`auth-form ${
-              dark === 'dark'
+              <label className={`auth-form ${dark === 'dark'
                 ? ' text-yellow-400'
                 : 'bg-white text-gray-700'
-            }`}>Password:</label>
+                }`}>Password:</label>
               <div className="password-wrapper">
                 <input
                   type={showPassword.password ? 'text' : 'password'}
@@ -186,11 +179,10 @@ const AuthPage = () => {
                   value={formData.password}
                   onChange={handleChange}
                   required
-                  className={`${
-                    dark === 'dark'
-                      ? ' text-yellow-400'
-                      : 'bg-white text-gray-900'
-                  }`}
+                  className={`${dark === 'dark'
+                    ? ' text-yellow-400'
+                    : 'bg-white text-gray-900'
+                    }`}
                 />
                 <button
                   type="button"
@@ -211,11 +203,10 @@ const AuthPage = () => {
             <div className="register-button">
               <button
                 type="submit"
-                className={`auth-button ${
-                  dark === 'dark'
-                    ? 'bg-yellow-400 text-gray-900 hover:bg-yellow-500'
-                    : 'bg-blue-600 text-white hover:bg-blue-700'
-                }`}
+                className={`auth-button ${dark === 'dark'
+                  ? 'bg-yellow-400 text-gray-900 hover:bg-yellow-500'
+                  : 'bg-blue-600 text-white hover:bg-blue-700'
+                  }`}
               >
                 Login
               </button>
@@ -223,9 +214,8 @@ const AuthPage = () => {
 
             <Link
               to="/register"
-              className={`toggle-auth-button text-center ${
-                dark === 'dark' ? 'text-yellow-400' : 'text-blue-600'
-              }`}
+              className={`toggle-auth-button text-center ${dark === 'dark' ? 'text-yellow-400' : 'text-blue-600'
+                }`}
             >
               Don't have an account? Register
             </Link>
