+++ client/src/pages/TermsAndConditions.jsx
